Move decleration of AlbumInfo
